+++ apps/web/src/app/index/magazine/list/page.tsx
@@ -1,14 +1,14 @@
 import { LoungeAndMagazineTab } from '@/components/common/lounge-and-magazine-tab'
 import { SideDrawer } from '@/components/common/side-drawer'
-import { LoungeTitle } from '@/components/lounge/lounge-magazine-title'
-import { useGetToken } from '@/hooks/use-get-token'
+import { MagazineTitle } from '@/components/magazine/magazine-title'
+import { useUpdateToken } from '@/hooks/use-update-token'
 import { Portal, usePortal } from '@/hooks/use-portal'
 import { MagazineCategoryKey } from '@/utils/query-key'
 import { AppScreen } from '@stackflow/plugin-basic-ui'
 import { cn } from '@yaksok/utils'
 
 export default function MagazineListPage() {
-  useGetToken()
+  useUpdateToken()
   const { portalRef, isOpen, setIsOpen } = usePortal()
 
   return (
@@ -49,7 +49,7 @@ const MagazineHeader = () => {
   return (
     <header className={cn('sticky top-[-248px] w-full bg-black')}>
       <div className="flex h-[248px] flex-col px-5 pt-25 pb-7.5 transition-all duration-500">
-        <LoungeTitle
+        <MagazineTitle
           titleClassName="text-gray07 text-subhead2 mt-4 mb-5"
           instagramClassName="text-body2 text-gray05"
         />
