+++ apps/web/src/app/index/page.tsx
@@ -4,12 +4,12 @@ import Counsel from '@/components/main/counsel'
 import MagazineCarousel from '@/components/main/magazine-carousel'
 import { MainCheck } from '@/components/main/main-check'
 import { MainLounge } from '@/components/main/main-lounge'
-import { useGetToken } from '@/hooks/use-get-token'
+import { useUpdateToken } from '@/hooks/use-update-token'
 import { AppScreen } from '@stackflow/plugin-basic-ui'
 import { Footer, Search } from '@yaksok/ui'
 
 export default function MainPage() {
-  useGetToken()
+  useUpdateToken()
 
   return (
     <AppScreen
