+++ apps/web/src/components/common/lounge-and-magazine-tab.tsx
@@ -10,7 +10,7 @@ import { PathType } from '@yaksok/api'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@yaksok/ui/tabs'
 import { MagazineListCard } from '@yaksok/ui'
 import { Suspense, useState } from 'react'
-import LoungeMagazineSelect from '../lounge/lounge-magazine-select'
+import LoungeMagazineSelect from './lounge-magazine-select'
 import { Magazine } from '@yaksok/api/boardMagazineType'
 import { useInView } from 'react-intersection-observer'
 import { useFlow } from '@/utils/stackflow'
