+++ apps/web/src/components/magazine/magazine-detail.tsx
@@ -1,0 +1,26 @@
+import { useHttpQuery } from '@/hooks/tanstak/use-http-query'
+import { useGetToken } from '@/hooks/use-get-token'
+import { MagazineDetail } from '@yaksok/api/boardMagazineType'
+
+export type MagazineDetailProps = {
+  id: string
+}
+export default function MagazineDetailShow({ id }: MagazineDetailProps) {
+  const token = useGetToken()
+  const result = useHttpQuery<undefined, MagazineDetail>(
+    ['magazine', id],
+    '/api/post/magazine/{postId}',
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+      params: {
+        postId: id,
+      },
+    }
+  )
+
+  console.log(result.data)
+
+  return <div>MagazineDetail</div>
+}
