+++ apps/web/src/components/magazine/magazine-yaksok-subscription.tsx
@@ -1,6 +1,6 @@
 import { UnderConstruction } from '../common'
 
-export default function LoungeYaksokSubscription() {
+export default function MagazineYaksokSubscription() {
   // 구독 프로그램이 존재하는 경우와 그렇지 않은 경우 분기처리 필요
   return (
     <div className="mt-14 flex flex-col gap-4">
